backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
pymongo==4.6.1
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-api-python-client==2.110.0
python-docx==1.1.0
PyPDF2==3.0.1
openpyxl==3.1.2
python-dotenv==1.0.0
python-dateutil==2.8.2
requests==2.31.0
Werkzeug==3.0.1
reportlab==4.0.7
pytesseract==0.3.10
Pillow==10.1.0
